+++ template/base_feature/src/main/java/ru/surfstudio/standard/ui/placeholder/loadstate/presentation/EmptyLoadStatePresentation.kt
@@ -1,15 +1,14 @@
 package ru.surfstudio.standard.ui.placeholder.loadstate.presentation
 
 import android.view.LayoutInflater
-import android.view.View
 import android.widget.TextView
 import androidx.annotation.StringRes
-import kotlinx.android.synthetic.main.layout_state_empty.view.*
 import ru.surfstudio.android.core.mvp.loadstate.SimpleLoadStatePresentation
 import ru.surfstudio.android.custom.view.placeholder.PlaceHolderViewContainer
 import ru.surfstudio.android.custom.view.placeholder.setClickableAndFocusable
 import ru.surfstudio.standard.ui.placeholder.loadstate.state.EmptyLoadState
 import ru.surfstudio.android.template.base_feature.R
+import ru.surfstudio.android.template.base_feature.databinding.LayoutStateEmptyBinding
 
 /**
  * Представление состояния EmptyLoadState, с картинкой, тайтлом, сабтайтлом и кнопкой
@@ -18,28 +17,21 @@ class EmptyLoadStatePresentation(
         private val placeHolder: PlaceHolderViewContainer
 ) : SimpleLoadStatePresentation<EmptyLoadState>() {
 
+    private val binding = LayoutStateEmptyBinding.inflate(LayoutInflater.from(placeHolder.context))
+
     @StringRes
     var messageTextRes: Int = R.string.state_empty_text
 
-    private lateinit var messageView: TextView
-
-    private val view: View by lazy {
-        LayoutInflater.from(placeHolder.context)
-                .inflate(R.layout.layout_state_empty, placeHolder, false)
-                .apply { messageView = empty_load_state_tv }
-    }
-
     override fun showState(state: EmptyLoadState) {
-        initViews(view)
-
+        initViews()
         with(placeHolder) {
-            changeViewTo(view)
+            changeViewTo(binding.root)
             setClickableAndFocusable(true)
             show()
         }
     }
 
-    private fun initViews(view: View) {
-        messageView.text = view.context.getString(messageTextRes)
+    private fun initViews() {
+        binding.emptyLoadStateTv.text = binding.root.context.getString(messageTextRes)
     }
 }