+++ template/base_feature/src/main/java/ru/surfstudio/standard/ui/screen_modules/FragmentScreenModule.kt
@@ -1,6 +1,7 @@
 package ru.surfstudio.standard.ui.screen_modules
 
 import android.content.Context
+import android.content.SharedPreferences
 import dagger.Module
 import dagger.Provides
 import ru.surfstudio.android.core.mvp.scope.FragmentViewPersistentScope
@@ -8,6 +9,7 @@ import ru.surfstudio.android.core.ui.event.ScreenEventDelegateManager
 import ru.surfstudio.android.core.ui.navigation.activity.navigator.ActivityNavigator
 import ru.surfstudio.android.core.ui.navigation.activity.navigator.ActivityNavigatorForFragment
 import ru.surfstudio.android.core.ui.navigation.feature.installer.SplitFeatureInstaller
+import ru.surfstudio.android.core.ui.permission.PermissionManager
 import ru.surfstudio.android.core.ui.provider.ActivityProvider
 import ru.surfstudio.android.core.ui.provider.FragmentProvider
 import ru.surfstudio.android.core.ui.scope.ScreenPersistentScope
@@ -18,8 +20,12 @@ import ru.surfstudio.android.message.DefaultMessageController
 import ru.surfstudio.android.message.MessageController
 import ru.surfstudio.android.mvp.dialog.navigation.navigator.DialogNavigator
 import ru.surfstudio.android.mvp.dialog.navigation.navigator.DialogNavigatorForFragment
+import ru.surfstudio.android.shared.pref.NO_BACKUP_SHARED_PREF
 import ru.surfstudio.android.template.base_feature.BuildConfig
 import ru.surfstudio.standard.ui.error.ErrorHandlerModule
+import ru.surfstudio.standard.v_message_controller_top.IconMessageController
+import ru.surfstudio.standard.v_message_controller_top.TopSnackIconMessageController
+import javax.inject.Named
 
 @Module(includes = [ErrorHandlerModule::class])
 class FragmentScreenModule(private val persistentScope: FragmentViewPersistentScope) : ScreenModule() {
@@ -49,6 +55,12 @@ class FragmentScreenModule(private val persistentScope: FragmentViewPersistentSc
             fragmentProvider: FragmentProvider
     ): MessageController {
         return DefaultMessageController(activityProvider, fragmentProvider)
+    }
+
+    @Provides
+    @PerScreen
+    internal fun provideTopMessageController(activityProvider: ActivityProvider): IconMessageController {
+        return TopSnackIconMessageController(activityProvider)
     }
 
     @Provides